+++ packages/webpack-config-terra/CHANGELOG.md
@@ -4,6 +4,9 @@
 
 * Added
   * Added `browser` to `resolve.mainFields` in webpack module list.
+  
+* Changed
+  * Passively consume @cerner/terra-aggregate-translations package.
 
 ## 1.2.0 - (January 5, 2021)